moved vector utils to WMS and applied to mapml features
